Switch: Renamed SDL_switchaudioout into SDL_switchaudio
As AudioRender support had been removed, the name of old module can be re-used normally.
